python_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
tensorflow==2.20.0
tf-keras==2.20.1
pillow==12.0.0
python-multipart==0.0.20

